# docs/api.yaml
openapi: 3.0.0
info:
  title: 趣搭星球 API
  version: 1.0.0
  description: 趣搭星球 - 基于兴趣的线下活动搭子匹配平台（支持单向关注 + 双向好友双轨制）
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
paths:
...
  /users/me:
    get:
      tags: [用户]
      summary: 获取当前登录用户信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      tags: [用户]
      summary: 编辑当前用户资料
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 更新成功
  /users/{id}:
    get:
      tags: [用户]
      summary: 获取指定用户的主页信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserHomepage'
  /users/location:
    post:
      tags: [用户]
      summary: 上报当前用户地理位置
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: 上报成功

  # ==================== 关注模块（单向） ====================
  /users/{id}/follow:
    post:
      tags: [关注]
      summary: 关注/取消关注用户（单向，无需对方同意）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - follow
              properties:
                follow:
                  type: boolean
                  description: true=关注，false=取消关注
      responses:
        '200':
          description: 操作成功
  /users/{id}/followers:
    get:
      tags: [关注]
      summary: 获取指定用户的粉丝列表
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResult'
  /users/{id}/following:
    get:
      tags: [关注]
      summary: 获取指定用户的关注列表
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResult'

  # ==================== 好友模块（双向） ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PageResult:
      type: object
      properties:
        total:
          type: integer
        list:
          type: array
          items: {}

    # 认证相关
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        tags:
          type: string
        followingCount:
          type: integer
          description: 关注数
        followerCount:
          type: integer
          description: 粉丝数
        friendCount:
          type: integer
          description: 双向好友数
        averageScore:
          type: number
          format: float
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        avatar:
          type: string
        tags:
          type: string
        age:
          type: integer
        gender:
          type: integer
    UpdateLocationRequest:
      type: object
      required:
        - longitude
        - latitude
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    UserHomepage:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        tags:
          type: string
        activityCount:
          type: integer
          description: 活动数
        averageScore:
          type: number
          format: float
          description: 评分
        rank:
          type: integer
          description: 排名
        isFollowing:
          type: boolean
          description: 当前用户是否已关注TA
        isFriend:
          type: boolean
          description: 当前用户是否与TA是双向好友
        publicDiaries:
          type: array
          items:
            $ref: '#/components/schemas/DiarySummary'
        recentActivities:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'

    # 好友相关
...
    ActivitySummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        activityTime:
          type: string
          format: date-time
        location:
          type: string
...
    DiarySummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        coverImage:
          type: string
        createTime:
          type: string
          format: date-time
...